agente_ia_local/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
requests==2.32.4
mysql-connector-python==9.4.0
python-dotenv==1.1.1
sqlglot==30.0.3
